+++ vehicle/OVMS.V3/components/vehicle_vweup/src/vweup_obd.cpp
@@ -1079,14 +1079,14 @@ void OvmsVehicleVWeUp::IncomingPollReply(const OvmsPoller::poll_job_t &job, uint
 
     case VWUP_BAT_MGMT_SOH_HIST: // (znams) Testing the reply from the PID 74CC
 
-      if (PollReply.FromUint8Mod("VWUP_BAT_MGMT_SOH_HIST", value, static_cast<uint16_t>(4))) {
+      if (PollReply.FromUint8Mod("VWUP_BAT_MGMT_SOH_HIST", value, 4)) {
         int i;    //Number of cell pack
         int totalBytes = (vweup_modelyear > 2019) ? 560 : 680;
         int byteCounter = 4; // Starting position of the first byte
         std::vector<int> sohArray(totalBytes);
         std::string resultSoh = "";
          for(i = 0; i < totalBytes; i++){
-              PollReply.FromUint8Mod("VWUP_BAT_MGMT_SOH_HIST", value, static_cast<uint16_t>(byteCounter));
+              PollReply.FromUint8Mod("VWUP_BAT_MGMT_SOH_HIST", value, byteCounter);
               sohArray[i] = value;
               char buffer[12];
               snprintf(buffer, sizeof(buffer), "%d", sohArray[i]);
